handled api exceptions

+++ src/user/user.controller.ts
@@ -9,8 +9,10 @@ import {
   HttpCode,
   HttpStatus,
   NotFoundException,
+  BadRequestException,
 } from '@nestjs/common';
 import { UsersService } from './user.service';
+import { Types } from 'mongoose';
 import { CreateUserDto, UpdateUserDto } from './user.dto';
 
 @Controller('api/users')
@@ -24,6 +26,9 @@ export class UsersController {
 
   @Get(':id')
   async findById(@Param('id') id: string) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new BadRequestException('Bad ObjectId in Body');
+    }
     const user = await this.usersService.findById(id);
     if (!user) {
       throw new NotFoundException('User not found');
@@ -39,6 +44,9 @@ export class UsersController {
 
   @Put(':id')
   async update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new BadRequestException('Bad ObjectId in Body');
+    }
     const updatedUser = await this.usersService.update(id, updateUserDto);
     if (!updatedUser) {
       throw new NotFoundException('User not found');
@@ -48,6 +56,9 @@ export class UsersController {
 
   @Delete(':id')
   async delete(@Param('id') id: string) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new BadRequestException('Bad ObjectId in Body');
+    }
     const user = await this.usersService.delete(id);
     if (!user) {
       throw new NotFoundException('User not found');
